Move bands to own file
