+++ cropharvest/eo/__init__.py
@@ -1,4 +1,4 @@
-from .eo import EarthEngineExporter, STATIC_BANDS, DYNAMIC_BANDS
+from .eo import EarthEngineExporter
 
 
-__all__ = ["EarthEngineExporter", "STATIC_BANDS", "DYNAMIC_BANDS"]
+__all__ = ["EarthEngineExporter"]
