+++ cropharvest/eo/era5.py
@@ -1,11 +1,10 @@
 import ee
 from datetime import date, timedelta
 
+from cropharvest.bands import ERA5_BANDS as BANDS
 from .utils import date_to_string
 
 image_collection = "ECMWF/ERA5_LAND/MONTHLY"
-
-BANDS = ["temperature_2m", "total_precipitation"]
 
 
 def get_single_image(region: ee.Geometry, start_date: date, end_date: date) -> ee.Image:
